refactor: split integration tests into several modules
