+++ tests/integration/custom_transactions_tests.rs
@@ -1,0 +1,237 @@
+use std::collections::BTreeMap;
+
+use miden_client::{
+    client::{
+        accounts::{AccountStorageMode, AccountTemplate},
+        transactions::transaction_request::TransactionRequest,
+    },
+    store::AuthInfo,
+};
+use miden_objects::{
+    accounts::AccountId,
+    assembly::ProgramAst,
+    assets::{FungibleAsset, TokenSymbol},
+    crypto::rand::{FeltRng, RpoRandomCoin},
+    notes::{
+        Note, NoteAssets, NoteExecutionMode, NoteInputs, NoteMetadata, NoteRecipient, NoteTag,
+        NoteType,
+    },
+    Felt, Word,
+};
+use miden_tx::utils::Serializable;
+
+use super::common::*;
+
+// CUSTOM TRANSACTION REQUEST
+// ================================================================================================
+//
+// The following functions are for testing custom transaction code. What the test does is:
+//
+// - Create a custom tx that mints a custom note which checks that the note args are as expected
+//   (ie, a word of 4 felts that represent [9, 12, 18, 3])
+//
+// - Create another transaction that consumes this note with custom code. This custom code only
+//   asserts that the {asserted_value} parameter is 0. To test this we first execute with
+//   an incorrect value passed in, and after that we try again with the correct value.
+//
+// Because it's currently not possible to create/consume notes without assets, the P2ID code
+// is used as the base for the note code.
+#[tokio::test]
+async fn test_transaction_request() {
+    let mut client = create_test_client();
+    wait_for_node(&mut client).await;
+
+    let account_template = AccountTemplate::BasicWallet {
+        mutable_code: false,
+        storage_mode: AccountStorageMode::Local,
+    };
+
+    client.sync_state().await.unwrap();
+    // Insert Account
+    let (regular_account, _seed) = client.new_account(account_template).unwrap();
+
+    let account_template = AccountTemplate::FungibleFaucet {
+        token_symbol: TokenSymbol::new("TEST").unwrap(),
+        decimals: 5u8,
+        max_supply: 10_000u64,
+        storage_mode: AccountStorageMode::Local,
+    };
+    let (fungible_faucet, _seed) = client.new_account(account_template).unwrap();
+
+    // Execute mint transaction in order to create custom note
+    let note = mint_custom_note(&mut client, fungible_faucet.id(), regular_account.id()).await;
+
+    client.sync_state().await.unwrap();
+
+    // Prepare transaction
+
+    // If these args were to be modified, the transaction would fail because the note code expects
+    // these exact arguments
+    let note_args = [[Felt::new(9), Felt::new(12), Felt::new(18), Felt::new(3)]];
+
+    let note_args_map = BTreeMap::from([(note.id(), Some(note_args[0]))]);
+
+    let code = "
+        use.miden::contracts::auth::basic->auth_tx
+        use.miden::kernels::tx::prologue
+        use.miden::kernels::tx::memory
+
+        begin
+            push.0 push.{asserted_value}
+            # => [0, {asserted_value}]
+            assert_eq
+
+            call.auth_tx::auth_tx_rpo_falcon512
+        end
+        ";
+
+    // FAILURE ATTEMPT
+
+    let failure_code = code.replace("{asserted_value}", "1");
+    let program = ProgramAst::parse(&failure_code).unwrap();
+
+    let tx_script = {
+        let account_auth = client.get_account_auth(regular_account.id()).unwrap();
+        let (pubkey_input, advice_map): (Word, Vec<Felt>) = match account_auth {
+            AuthInfo::RpoFalcon512(key) => (
+                key.public_key().into(),
+                key.to_bytes().iter().map(|a| Felt::new(*a as u64)).collect::<Vec<Felt>>(),
+            ),
+        };
+
+        let script_inputs = vec![(pubkey_input, advice_map)];
+        client.compile_tx_script(program, script_inputs, vec![]).unwrap()
+    };
+
+    let transaction_request = TransactionRequest::new(
+        regular_account.id(),
+        note_args_map.clone(),
+        vec![],
+        Some(tx_script),
+    );
+
+    // This fails becuase of {asserted_value} having the incorrect number passed in
+    assert!(client.new_transaction(transaction_request).is_err());
+
+    // SUCCESS EXECUTION
+
+    let success_code = code.replace("{asserted_value}", "0");
+    let program = ProgramAst::parse(&success_code).unwrap();
+
+    let tx_script = {
+        let account_auth = client.get_account_auth(regular_account.id()).unwrap();
+        let (pubkey_input, advice_map): (Word, Vec<Felt>) = match account_auth {
+            AuthInfo::RpoFalcon512(key) => (
+                key.public_key().into(),
+                key.to_bytes().iter().map(|a| Felt::new(*a as u64)).collect::<Vec<Felt>>(),
+            ),
+        };
+
+        let script_inputs = vec![(pubkey_input, advice_map)];
+        client.compile_tx_script(program, script_inputs, vec![]).unwrap()
+    };
+
+    let transaction_request =
+        TransactionRequest::new(regular_account.id(), note_args_map, vec![], Some(tx_script));
+
+    execute_tx_and_sync(&mut client, transaction_request).await;
+
+    client.sync_state().await.unwrap();
+}
+
+async fn mint_custom_note(
+    client: &mut TestClient,
+    faucet_account_id: AccountId,
+    target_account_id: AccountId,
+) -> Note {
+    // Prepare transaction
+    let mut random_coin = RpoRandomCoin::new(Default::default());
+    let note = create_custom_note(client, faucet_account_id, target_account_id, &mut random_coin);
+
+    let recipient = note
+        .recipient_digest()
+        .iter()
+        .map(|x| x.as_int().to_string())
+        .collect::<Vec<_>>()
+        .join(".");
+
+    let note_tag = note.metadata().tag().inner();
+
+    let code = "
+    use.miden::contracts::faucets::basic_fungible->faucet
+    use.miden::contracts::auth::basic->auth_tx
+    
+    begin
+        push.{recipient}
+        push.{note_type}
+        push.{tag}
+        push.{amount}
+        call.faucet::distribute
+    
+        call.auth_tx::auth_tx_rpo_falcon512
+        dropw dropw
+    end
+    "
+    .replace("{recipient}", &recipient)
+    .replace("{note_type}", &Felt::new(NoteType::OffChain as u64).to_string())
+    .replace("{tag}", &Felt::new(note_tag.into()).to_string())
+    .replace("{amount}", &Felt::new(10).to_string());
+
+    let program = ProgramAst::parse(&code).unwrap();
+
+    let tx_script = {
+        let account_auth = client.get_account_auth(faucet_account_id).unwrap();
+        let (pubkey_input, advice_map): (Word, Vec<Felt>) = match account_auth {
+            AuthInfo::RpoFalcon512(key) => (
+                key.public_key().into(),
+                key.to_bytes().iter().map(|a| Felt::new(*a as u64)).collect::<Vec<Felt>>(),
+            ),
+        };
+
+        let script_inputs = vec![(pubkey_input, advice_map)];
+        client.compile_tx_script(program, script_inputs, vec![]).unwrap()
+    };
+
+    let transaction_request = TransactionRequest::new(
+        faucet_account_id,
+        BTreeMap::new(),
+        vec![note.clone()],
+        Some(tx_script),
+    );
+
+    let _ = execute_tx_and_sync(client, transaction_request).await;
+    note
+}
+
+fn create_custom_note(
+    client: &TestClient,
+    faucet_account_id: AccountId,
+    target_account_id: AccountId,
+    rng: &mut RpoRandomCoin,
+) -> Note {
+    let expected_note_arg = [Felt::new(9), Felt::new(12), Felt::new(18), Felt::new(3)]
+        .iter()
+        .map(|x| x.to_string())
+        .collect::<Vec<_>>()
+        .join(".");
+
+    let note_script =
+        include_str!("asm/custom_p2id.masm").replace("{expected_note_arg}", &expected_note_arg);
+    let note_script = ProgramAst::parse(&note_script).unwrap();
+    let note_script = client.compile_note_script(note_script, vec![]).unwrap();
+
+    let inputs = NoteInputs::new(vec![target_account_id.into()]).unwrap();
+    let serial_num = rng.draw_word();
+    let note_metadata = NoteMetadata::new(
+        faucet_account_id,
+        NoteType::OffChain,
+        NoteTag::from_account_id(target_account_id, NoteExecutionMode::Local)
+            .unwrap(),
+        Default::default(),
+    )
+    .unwrap();
+    let note_assets =
+        NoteAssets::new(vec![FungibleAsset::new(faucet_account_id, 10).unwrap().into()]).unwrap();
+    let note_recipient = NoteRecipient::new(serial_num, note_script, inputs);
+    Note::new(note_assets, note_metadata, note_recipient)
+}
